refactor: auth form shared code
move to appropriate folder
