+++ src/credentials/vue/authForms/baseAuth.vue
@@ -1,5 +1,6 @@
 <script lang="ts">
 import { defineComponent } from 'vue'
+import { AuthFormId, AuthStatus } from './shared.vue'
 
 export default defineComponent({
     emits: ['auth-connection-updated'],
@@ -10,22 +11,9 @@ export default defineComponent({
     },
 })
 
-export interface AuthStatus {
-    /**
-     * Returns true if the auth is successfully connected.
-     */
-    isAuthConnected(): Promise<boolean>
-}
-
 export class UnimplementedAuthStatus implements AuthStatus {
     isAuthConnected(): Promise<boolean> {
         return Promise.resolve(false)
     }
 }
-
-export const authForms = {
-    CREDENTIALS: 'CREDENTIALS',
-} as const
-
-export type AuthFormId = (typeof authForms)[keyof typeof authForms]
 </script>
